Refactor for multiple payment methods.
